feat(Unity): add GameObjectAdapter and refactor input adapters
- Introduce GameObjectAdapter to bridge Lunar and Unity GameObjects
- Move InputActionsAdapter into InputAdapter.cs file
- Remove separate InputActionsAdapter.cs file
- Update meta file for renamed GameObjectAdapter
