+++ Runtime/GameObjectAdapter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0362c5da299609736b58dffaca9bbe76
+guid: c9383aed2c0cdc9789fa0af3c16524d5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
